pyproject.toml
[tool.poetry]
name = "git-co-evg-base"
version = "3.1.0"
description = "Find a good commit to base your work on"
authors = ["David Bradford <david.bradford@mongodb.com>"]
readme = "README.md"
repository = "https://github.com/evergreen-ci/git-co-evg-base"
packages = [
    { include = "goodbase", from = "src"},
]

[tool.poetry.scripts]
git-co-evg-base = "goodbase.goodbase_cli:main"

[tool.poetry.dependencies]
python = "^3.7.1"
click = "^8"
"evergreen.py" = "^3.2.7"
plumbum = "^1.7.0"
Inject = "^4.3.1"
structlog = "^21.1.0"
PyYAML = "^5.4.1"
xdg = "^5.1.1"
pydantic = "^1.8.2"
rich = "^10.9.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
black = "^22.3.0"
pytest-black = "^0.3"
pytest-cov = "^2.8"
pytest-flake8 = "^1.0"
pytest-mypy = "^0.8"
mypy = "^0.910"
pytest-pydocstyle = "^2.0"
pre-commit = "^2.6"
pytest-isort = "^2.0"
flake8-bugbear = "^21.4"
types-PyYAML = "^5.4.10"
types-requests = "^2.27.7"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 100

[tool.pytest.ini_options]
flake8-ignore = "W605 W503 W291 E203 E501 F821"
addopts = "--flake8 --black --mypy --isort --pydocstyle"
testpaths = [
    "src",
    "tests",
]

[tool.mypy]
ignore_missing_imports = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
disallow_untyped_calls = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "tests.*",
]
ignore_errors = true

[tool.pydocstyle]
ignore = "D104,D203,D212,D301,D407,D412"
match = "(?!(?:test_|conftest)).*\\.py"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
